Add support for ssd1306 128x64 OLED Display
Included olikraus u8g2 library to driver the display
and a set of primitives to make it possible for the u8g2
to handle the i2c interface of the lpc.

It was necessary to have an operational i2c, for which,
at least temporarily, I will be using the LPCOpen driver
for it. (The other option was software i2c, which I'll pass).

Also added some delay() functions, they're useful
for initializing things like the LCD and testing stuff.

Last but not least, minor tweaks and improvements
in things like gpio access macros.